+++ client/browser/src/browser-extension/scripts/contentPage.main.ts
@@ -7,18 +7,16 @@ import { setLinkComponent, AnchorLink } from '@sourcegraph/shared/src/components
 
 import { determineCodeHost } from '../../shared/code-hosts/shared/codeHost'
 import { injectCodeIntelligence } from '../../shared/code-hosts/shared/inject'
+import { logger } from '../../shared/code-hosts/shared/util/logger'
 import {
-    checkIsSourcegraph,
     EXTENSION_MARKER_ID,
     injectExtensionMarker,
     NATIVE_INTEGRATION_ACTIVATED,
-    signalBrowserExtensionInstalled,
 } from '../../shared/code-hosts/sourcegraph/inject'
 import { initSentry } from '../../shared/sentry'
-import { DEFAULT_SOURCEGRAPH_URL, getAssetsURL } from '../../shared/util/context'
+import { CLOUD_SOURCEGRAPH_URL, getAssetsURL } from '../../shared/util/context'
 import { featureFlags } from '../../shared/util/featureFlags'
 import { assertEnvironment } from '../environmentAssertion'
-import { storage } from '../web-extension-api/storage'
 
 const subscriptions = new Subscription()
 window.addEventListener('unload', () => subscriptions.unsubscribe(), { once: true })
@@ -29,8 +27,6 @@ const codeHost = determineCodeHost()
 initSentry('content', codeHost?.type)
 
 setLinkComponent(AnchorLink)
-
-const IS_EXTENSION = true
 
 // Add style sheet and wait for it to load to avoid rendering unstyled elements (which causes an
 // annoying flash/jitter when the stylesheet loads shortly thereafter).
@@ -66,7 +62,7 @@ async function waitForStyleSheet(styleSheet: HTMLLinkElement): Promise<void> {
  * Main entry point into browser extension.
  */
 async function main(): Promise<void> {
-    console.log('Sourcegraph browser extension is running')
+    logger.info('Browser extension is running')
 
     // Make sure DOM is fully loaded
     if (document.readyState !== 'complete' && document.readyState !== 'interactive') {
@@ -81,7 +77,7 @@ async function main(): Promise<void> {
     // If the native integration was activated before the content script, we can
     // synchronously check for the presence of the extension marker.
     if (document.querySelector(`#${EXTENSION_MARKER_ID}`) !== null) {
-        console.log('Sourcegraph native integration is already running')
+        logger.info('Native integration is already running')
         return
     }
     // If the extension marker isn't present, inject it and listen for a custom event sent by the native
@@ -91,41 +87,28 @@ async function main(): Promise<void> {
         .pipe(first())
         .toPromise()
 
-    const items = await storage.sync.get()
-    const sourcegraphURL = items.sourcegraphURL || DEFAULT_SOURCEGRAPH_URL
-
-    const isSourcegraphServer = checkIsSourcegraph(sourcegraphURL)
-    if (isSourcegraphServer) {
-        signalBrowserExtensionInstalled()
-        return
-    }
-
     subscriptions.add(
-        await injectCodeIntelligence(
-            { sourcegraphURL, assetsURL: getAssetsURL(DEFAULT_SOURCEGRAPH_URL) },
-            IS_EXTENSION,
-            async function onCodeHostFound() {
-                const styleSheets = [
-                    {
-                        id: 'ext-style-sheet',
-                        path: 'css/style.bundle.css',
-                    },
-                    {
-                        id: 'ext-style-sheet-css-modules',
-                        path: 'css/inject.bundle.css',
-                    },
-                ]
+        await injectCodeIntelligence(getAssetsURL(CLOUD_SOURCEGRAPH_URL), true, async function onCodeHostFound() {
+            const styleSheets = [
+                {
+                    id: 'ext-style-sheet',
+                    path: 'css/style.bundle.css',
+                },
+                {
+                    id: 'ext-style-sheet-css-modules',
+                    path: 'css/inject.bundle.css',
+                },
+            ]
 
-                await Promise.all(styleSheets.map(loadStyleSheet).map(waitForStyleSheet))
-            }
-        )
+            await Promise.all(styleSheets.map(loadStyleSheet).map(waitForStyleSheet))
+        })
     )
 
     // Clean up susbscription if the native integration gets activated
     // later in the lifetime of the content script.
     await nativeIntegrationActivationEventReceived
-    console.log('Native integration activation event received')
+    logger.info('Native integration activation event received')
     subscriptions.unsubscribe()
 }
 
-main().catch(console.error.bind(console))
+main().catch(console.error)