+++ client/browser/src/shared/code-hosts/shared/inject.ts
@@ -1,10 +1,32 @@
 import { Observable, Subscription } from 'rxjs'
-import { startWith } from 'rxjs/operators'
+import { filter, startWith } from 'rxjs/operators'
 
-import { SourcegraphIntegrationURLs } from '../../platform/context'
+import { isCloudSupportedCodehost, SourcegraphUrlService } from '../../platform/sourcegraphUrlService'
 import { MutationRecordLike, observeMutations as defaultObserveMutations } from '../../util/dom'
 
 import { determineCodeHost, CodeHost, injectCodeIntelligenceToCodeHost, ObserveMutations } from './codeHost'
+import { RepoURLParseError } from './errors'
+import { logger } from './util/logger'
+
+function inject(codeHost: CodeHost, assetsURL: string, sourcegraphURL: string, isExtension: boolean): Subscription {
+    logger.info('Attaching code intelligence using', sourcegraphURL)
+
+    const observeMutations: ObserveMutations = codeHost.observeMutations || defaultObserveMutations
+    const mutations: Observable<MutationRecordLike[]> = observeMutations(document.body, {
+        childList: true,
+        subtree: true,
+    }).pipe(startWith([{ addedNodes: [document.body], removedNodes: [] }]))
+
+    return injectCodeIntelligenceToCodeHost(
+        mutations,
+        codeHost,
+        {
+            assetsURL,
+            sourcegraphURL,
+        },
+        isExtension
+    )
+}
 
 /**
  * Checks if the current page is a known code host. If it is,
@@ -15,25 +37,38 @@ import { determineCodeHost, CodeHost, injectCodeIntelligenceToCodeHost, ObserveM
  * such work on unsupported websites
  */
 export async function injectCodeIntelligence(
-    urls: SourcegraphIntegrationURLs,
+    assetsURL: string,
     isExtension: boolean,
-    onCodeHostFound?: (codeHost: CodeHost) => Promise<void>
+    onCodeHostFound?: (codeHost: CodeHost) => Promise<void>,
+    overrideSourcegraphURL?: string
 ): Promise<Subscription> {
-    const subscriptions = new Subscription()
-    const codeHost = determineCodeHost(urls.sourcegraphURL)
-    if (codeHost) {
-        console.log('Sourcegraph: Detected code host:', codeHost.type)
-
-        if (onCodeHostFound) {
-            await onCodeHostFound(codeHost)
-        }
-
-        const observeMutations: ObserveMutations = codeHost.observeMutations || defaultObserveMutations
-        const mutations: Observable<MutationRecordLike[]> = observeMutations(document.body, {
-            childList: true,
-            subtree: true,
-        }).pipe(startWith([{ addedNodes: [document.body], removedNodes: [] }]))
-        subscriptions.add(injectCodeIntelligenceToCodeHost(mutations, codeHost, urls, isExtension))
+    const codeHost = determineCodeHost()
+    if (!codeHost) {
+        return new Subscription()
     }
-    return subscriptions
+
+    if (onCodeHostFound) {
+        await onCodeHostFound(codeHost)
+    }
+
+    if (overrideSourcegraphURL) {
+        return inject(codeHost, assetsURL, overrideSourcegraphURL, isExtension)
+    }
+    logger.info(`Detected: codehost="${codeHost.type}"`)
+
+    try {
+        const { rawRepoName } = codeHost.getContext?.() || {}
+        logger.info(`Detected repository="${rawRepoName ?? ''}"`)
+        if (rawRepoName) {
+            await SourcegraphUrlService.use(rawRepoName)
+        }
+    } catch (error) {
+        if (!(error instanceof RepoURLParseError)) {
+            throw error
+        }
+    }
+
+    return SourcegraphUrlService.observe(isExtension)
+        .pipe(filter(isCloudSupportedCodehost))
+        .subscribe(sourcegraphURL => inject(codeHost, assetsURL, sourcegraphURL, isExtension))
 }