+++ client/web/src/Layout.tsx
@@ -127,7 +127,6 @@ export interface LayoutProps
     globbing: boolean
     showMultilineSearchConsole: boolean
     showSearchNotebook: boolean
-    showQueryBuilder: boolean
     isSourcegraphDotCom: boolean
     fetchSavedSearches: () => Observable<GQL.ISavedSearch[]>
     children?: never