+++ client/web/src/integration/search-contexts.test.ts
@@ -8,7 +8,7 @@ import { ISearchContext } from '@sourcegraph/shared/src/graphql/schema'
 import { Driver, createDriverForTest } from '@sourcegraph/shared/src/testing/driver'
 import { afterEachSaveScreenshotIfFailed } from '@sourcegraph/shared/src/testing/screenshotReporter'
 
-import { RepoGroupsResult, SearchSuggestionsResult, WebGraphQlOperations } from '../graphql-operations'
+import { WebGraphQlOperations } from '../graphql-operations'
 
 import { WebIntegrationTestContext, createWebIntegrationTestContext } from './context'
 import { createRepositoryRedirectResult } from './graphQlResponseHelpers'
@@ -17,14 +17,6 @@ import { createJsContext, siteGQLID, siteID } from './jscontext'
 
 const commonSearchGraphQLResults: Partial<WebGraphQlOperations & SharedGraphQlOperations> = {
     ...commonWebGraphQlResults,
-    SearchSuggestions: (): SearchSuggestionsResult => ({
-        search: {
-            suggestions: [],
-        },
-    }),
-    RepoGroups: (): RepoGroupsResult => ({
-        repoGroups: [],
-    }),
     ConvertVersionContextToSearchContext: ({ name }) => ({
         convertVersionContextToSearchContext: { id: `id${name}`, spec: name },
     }),
@@ -216,10 +208,10 @@ describe('Search contexts', () => {
         await driver.page.waitForSelector('.test-convert-version-context-btn', { visible: true })
         await driver.page.click('.test-convert-version-context-btn')
 
-        await driver.page.waitForSelector('.convert-version-context-node .text-success')
+        await driver.page.waitForSelector('[data-testid="convert-version-context-node"] .text-success')
 
         const successText = await driver.page.evaluate(
-            () => document.querySelector('.convert-version-context-node .text-success')?.textContent
+            () => document.querySelector('[data-testid="convert-version-context-node"] .text-success')?.textContent
         )
         expect(successText).toBe('Version context successfully converted.')
     })