+++ cmd/frontend/graphqlbackend/insights.go
@@ -20,9 +20,13 @@ type InsightsResolver interface {
 
 	// Mutations
 	CreateInsightsDashboard(ctx context.Context, args *CreateInsightsDashboardArgs) (InsightsDashboardPayloadResolver, error)
+	UpdateInsightsDashboard(ctx context.Context, args *UpdateInsightsDashboardArgs) (InsightsDashboardPayloadResolver, error)
 	DeleteInsightsDashboard(ctx context.Context, args *DeleteInsightsDashboardArgs) (*EmptyResponse, error)
 	RemoveInsightViewFromDashboard(ctx context.Context, args *RemoveInsightViewFromDashboardArgs) (InsightsDashboardPayloadResolver, error)
 	AddInsightViewToDashboard(ctx context.Context, args *AddInsightViewToDashboardArgs) (InsightsDashboardPayloadResolver, error)
+
+	// Admin Management Mutations
+	UpdateInsightSeries(ctx context.Context, args *UpdateInsightSeriesArgs) (InsightSeriesMetadataPayloadResolver, error)
 }
 
 type InsightsArgs struct {
@@ -100,6 +104,16 @@ type CreateInsightsDashboardInput struct {
 	Grants InsightsPermissionGrants
 }
 
+type UpdateInsightsDashboardArgs struct {
+	Id    graphql.ID
+	Input UpdateInsightsDashboardInput
+}
+
+type UpdateInsightsDashboardInput struct {
+	Title  *string
+	Grants *InsightsPermissionGrants
+}
+
 type InsightsPermissionGrants struct {
 	Users         *[]graphql.ID
 	Organizations *[]graphql.ID
@@ -143,4 +157,23 @@ type RemoveInsightViewFromDashboardArgs struct {
 type RemoveInsightViewFromDashboardInput struct {
 	InsightViewID graphql.ID
 	DashboardID   graphql.ID
+}
+
+type UpdateInsightSeriesArgs struct {
+	Input UpdateInsightSeriesInput
+}
+
+type UpdateInsightSeriesInput struct {
+	SeriesId string
+	Enabled  *bool
+}
+
+type InsightSeriesMetadataResolver interface {
+	SeriesId(ctx context.Context) (string, error)
+	Query(ctx context.Context) (string, error)
+	Enabled(ctx context.Context) (bool, error)
+}
+
+type InsightSeriesMetadataPayloadResolver interface {
+	Series(ctx context.Context) InsightSeriesMetadataResolver
 }